requirements.txt
# ── Budget pipeline ───────────────────────────────────────────────────────────
pandas>=2.2.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.1.2
openpyxl>=3.1.0
openai>=1.13.0

# ── Reform pipeline ───────────────────────────────────────────────────────────
anthropic>=0.25.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
tenacity>=8.2.0
httpx>=0.27.0
playwright>=1.40.0
